included proximity sensor in meeting activity, updated sent request format for database
