+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
             android:exported="false" />
         <activity
             android:name=".activity.InMeetingActivity"
-            android:exported="false" />
+            android:exported="true" />
         <activity
             android:name=".MainActivity"
             android:exported="true">
@@ -35,9 +35,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".activity.UserProximityInput"
-            android:exported="true" />
     </application>
 
 </manifest>
