+++ app/src/main/java/com/group7/meetr/activity/InMeetingActivity.java
@@ -2,21 +2,41 @@ package com.group7.meetr.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
+import android.hardware.Sensor;
+import android.hardware.SensorEvent;
+import android.hardware.SensorEventListener;
+import android.hardware.SensorManager;
 import android.os.Bundle;
 
 import com.group7.meetr.R;
+import com.group7.meetr.data.remote.SessionHandler;
 
 import android.os.Vibrator;
 import android.view.View;
 import android.widget.Button;
+import android.widget.Toast;
 
-public class InMeetingActivity extends AppCompatActivity {
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
+public class InMeetingActivity extends AppCompatActivity implements SensorEventListener {
+
+    private SensorManager sensorManager;
+    Sensor proximitySensor;
+    private SessionHandler sessionHandler;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_participant_view);
+
+        sensorManager = (SensorManager) getSystemService(Context.SENSOR_SERVICE);
+        proximitySensor = sensorManager.getDefaultSensor(Sensor.TYPE_PROXIMITY);
+        sensorManager.registerListener(this, proximitySensor, SensorManager.SENSOR_DELAY_NORMAL);
+
+        sessionHandler = new SessionHandler();
 
         Button vib = findViewById(R.id.buttonJoin);
         Vibrator vibr = (Vibrator) getSystemService(Context.VIBRATOR_SERVICE);
@@ -26,5 +46,48 @@ public class InMeetingActivity extends AppCompatActivity {
                 vibr.vibrate(400);
             }
         });
+    }
+    @Override
+    public void onSensorChanged(SensorEvent event) {
+        float distance = event.values[0];
+        long timestamp = System.currentTimeMillis();
+
+        if (distance <= 7) {
+            Toast.makeText(this, "Request Registered: " + timestamp, Toast.LENGTH_SHORT).show();
+            // Send handRaise flag and timestamp to sessionHandler
+            String sessionId = "7";
+            sessionHandler.sendProximityData(sessionId, true, timestamp);
+        }
+        else {
+            Toast.makeText(this, "Request Not Registered: " + timestamp, Toast.LENGTH_SHORT).show();
+        }
+    }
+
+    @Override
+    public void onAccuracyChanged(Sensor sensor, int accuracy) {
+
+    }
+
+    @Override
+    protected void onResume() {
+        // Register a listener for the sensor.
+        super.onResume();
+        sensorManager.registerListener(this, proximitySensor, SensorManager.SENSOR_DELAY_NORMAL);
+    }
+
+    @Override
+    protected void onPause() {
+        // Be sure to unregister the sensor when the activity pauses.
+        super.onPause();
+        sensorManager.unregisterListener(this);
+    }
+    @SuppressLint("SimpleDateFormat")
+    public static String getCurrentTimeStamp(){
+        try {
+            return new SimpleDateFormat("yyyy-MM-dd HH:mm:ss").format(new Date());
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
     }
 }
